move shuffleArray() into helper file, shuffle offers on offer pages, wrap h4 offer headings that are too long
